164-BE: Fix Cursor pagination with decimal amount

+++ Okane.Api/Features/Finances/Dtos/FinanceRecordCursorQueryParameters.cs
@@ -5,6 +5,6 @@ public class FinanceRecordCursorQueryParameters
     public int? CursorId { get; set; }
 
     // One of CursorAmount and CursorHappenedAt should be passed, but not both.
-    public int? CursorAmount { get; set; }
+    public decimal? CursorAmount { get; set; }
     public DateTime? CursorHappenedAt { get; set; }
 }
